refactor(mapping): separate BA solving from snapshot capture and writeback
From work authored in this worktree by Codex.

Bundle adjustment sat inside the map, so the map held solver policy and the
solvers held raw storage. Split along that seam: `map::ops::snapshot` owns
`BaSnapshot` — consistent inputs captured under the caller's lock, with no
window or solver policy — and `BaUpdate`, which carries only numeric
estimates and refreshed preintegrations back. Observations, landmark
lifetimes and other live structure are never copied back, so an in-flight
snapshot structurally cannot resurrect what was retired while it solved.
`mapping::bundle_adjustment` owns `run_initial_ba`, `run_local_ba` and
`run_local_inertial_ba`, selecting their problems from those immutable
inputs; correction applies the result and keeps the epoch and `culled`
guards.

`ImuFactor` moves to its own entity module beside `Keyframe` and
`MapPoint`, and the larger test modules move into sibling files, leaving
`mutation.rs` and the map root readable.

Behaviour-preserving across four configurations: MH_01 stereo 0.0255 m ATE
/ mono 0.0312 m, V1_01 stereo 0.0716 m / mono 0.0892 m, all within
run-to-run wobble of 7745bbe. 199 library tests, up from 193.
