--- crates/kornia-slam/src/mapping/map/ops/Untitled
@@ -1,1 +1,0 @@
-add_triangulated_points
