+++ crates/kornia-slam/src/mapping/mod.rs
@@ -1,10 +1,11 @@
 //! Mapping: the map itself, plus the algorithms and worker that operate on it.
 //!
 //! [`map`] owns the stored entities, structural queries, and the canonical
-//! mutation and correction operations. Growth, culling and bundle-adjustment
+//! mutation and correction operations. Growth, culling, bundle adjustment and
 //! scheduling are mapping algorithms layered on top; they read the map
 //! immutably and write only through its operations.
 
+pub mod bundle_adjustment;
 pub mod culling;
 pub mod growth;
 pub mod local_mapping;
